+++ assets/libs/tty/index.js
@@ -1,5 +1,9 @@
 if (FuseBox.isServer) {
     module.exports = global.require("tty");
 } else {
-    module.exports = {}
+    module.exports = {
+        isatty: function() {
+            return false;
+        }
+    }
 }