+++ src/plugins/SassPlugin.ts
@@ -1,4 +1,3 @@
-import * as path from "path";
 import { File } from "../File";
 import { WorkFlowContext } from "../WorkflowContext";
 import { Plugin } from "../WorkflowContext";
@@ -41,7 +40,7 @@ export class SassPluginClass implements Plugin {
         return new Promise((res, rej) => {
             return sass.render(options, (err, result) => {
                 if (err) { return rej(err); }
-                file.sourceMap = result.map.toString();
+                file.sourceMap = result.map && result.map.toString();
                 file.contents = result.css.toString();
                 return res(file.contents);
             });